refactor `submitScores`
